Suma1C pasado a float

+++ Clase1/Suma1C/main.c
@@ -3,15 +3,15 @@
 
 int main()
 {
-    int numero1;
-    int numero2;
-    int suma;
+    float numero1;
+    float numero2;
+    float suma;
     printf("Ingrese el primer numero: ");
-    scanf("%d" , &numero1);
+    scanf("%f" , &numero1);
     printf("Ingrese el segundo numero: ");
-    scanf("%d" , &numero2);
+    scanf("%f" , &numero2);
     suma = numero1 + numero2;
-    printf("El resultado de la suma es: %d" , suma);
+    printf("El resultado de la suma es: %f" , suma);
     system("pause");
     return 0;
 }
